third_party/blink/web_tests/fast/dom/shadow/shadow-root-direction.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id="target" dir="rtl">
</div>
<script>
description('Element direction details are accessible to shadow root');
var root = document.getElementById('target');
var shadowRoot = root.attachShadow({mode: 'open'});
shadowRoot.innerHTML = "<div id='target2'> hello! </div>"

shouldBeEqualToString('getComputedStyle(shadowRoot.getElementById("target2")).direction', 'rtl'); 
shouldBeEqualToString('getComputedStyle(root).direction', 'rtl');

</script>
