third_party/blink/web_tests/fast/dom/shadow/shadow-root-direction-expected.txt

Element direction details are accessible to shadow root

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(shadowRoot.getElementById("target2")).direction is "rtl"
PASS getComputedStyle(root).direction is "rtl"
PASS successfullyParsed is true

TEST COMPLETE
